adding final piece of OOP refactoring
