--- node-app/lib/api.js
@@ -1,3 +1,0 @@
-//some general purpose utilities
-var util = require('./util');
-exports.util = util;
