+++ node-app/lib/navRoutesInitializer.js
@@ -1,7 +1,7 @@
 var navMainRouter = require(process.cwd() + "/lib/navMainRouter.js"),
     navSignInRouter = require(process.cwd() + "/lib/navSignInRouter.js"),
     navRegistrationRouter = require(process.cwd() + "/lib/navRegistrationRouter.js"),
-    navnirmiteeApi = require(process.cwd() + "/lib/api.js");
+    navToysRouter = require(process.cwd() + "/lib/navToysRouter.js");
 
 
 module.exports = class navRoutesInitializer {
@@ -20,13 +20,9 @@ module.exports = class navRoutesInitializer {
             app.use('/', new navMainRouter().setup().getRouter());
             app.use('/', new navSignInRouter().setup().getRouter());
             app.use('/', new navRegistrationRouter().setup().getRouter());
-            app.use('/', require(process.cwd() + '/routes/temp.js'));	
+            app.use('/', require(process.cwd() + '/lib/temp.js'));	
             //app.use('/registration', require('./routes/regAndAuth/registration.js'));
-            app.use('/toys', 
-                    navnirmiteeApi.util.ensureVerified,
-                    navnirmiteeApi.util.ensureAuthenticated,
-                    navnirmiteeApi.util.isSessionAvailable,
-                    require(process.cwd() + '/routes/toys/toyDetail.js'));
+            app.use('/toys',new navToysRouter().setup().getRouter());
         
     }
 }
