--- node-app/lib_old/exceptions/navEmailVerificationException.js
@@ -1,9 +1,0 @@
-const NE = require('node-exceptions')
-class navEmailVerificationException extends NE.LogicalException {
-    constructor() {
-        super("Please verify email id...",400,"VERIFY_EMAIL");
-    }
-
-}
- 
-module.exports = navUserNotFoundException;
