--- node-app/lib_old/exceptions/navUserExistsException.js
@@ -1,9 +1,0 @@
-const NE = require('node-exceptions')
-class navUserExistsException extends NE.LogicalException {
-    constructor() {
-        super("User Exists",400,"DUPL_USER");
-    }
-
-}
- 
-module.exports = navUserExistsException;
