--- node-app/lib_old/exceptions/navUserNotFoundException.js
@@ -1,9 +1,0 @@
-const NE = require('node-exceptions')
-class navUserNotFoundException extends NE.LogicalException {
-    constructor() {
-        super("User not found",500,"INVALID_USER");
-    }
-
-}
- 
-module.exports = navUserNotFoundException;
